% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_clust.R
\name{em_clust}
\alias{em_clust}
\title{Clusters observations into k clusters using the EM algorithm}
\usage{
em_clust(dat, k)
}
\arguments{
\item{dat}{A dataframe}

\item{k}{Number of clusters}
}
\value{
A vector of cluster assignments
}
\description{
Clusters observations into k clusters using the EM algorithm
}
